Move multi_agent_v2 feature config
